+++ app/src/main/java/com/milesilac/classreadingstats/persistence/model/SectionEntity.kt
@@ -7,7 +7,7 @@ import com.milesilac.classreadingstats.model.GradeLevel
 
 @Entity(tableName = "sections")
 data class SectionEntity(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "section_room_id") val sectionRoomId: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "section_room_id") val sectionRoomId: Long = 0,
     @ColumnInfo(name = "grade_level") val gradeLevel: GradeLevel,
     @ColumnInfo(name = "section_name") val sectionName: String,
 )
